refactor(racing): controller와 view 분리
