+++ src/main/java/com/racing/model/Game.java
@@ -1,8 +1,9 @@
-package com.racing;
+package com.racing.model;
+
+import com.racing.model.Car;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class Game {
