+++ src/test/java/com/racing/CarTest.java
@@ -1,5 +1,6 @@
 package com.racing;
 
+import com.racing.model.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
